+++ wikimedia/json-codec/src/JsonStdClassCodec.php
@@ -2,31 +2,19 @@
 declare( strict_types=1 );
 
 /**
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- * http://www.gnu.org/copyleft/gpl.html
- *
+ * @license GPL-2.0-or-later
  * @file
  */
 
 namespace Wikimedia\JsonCodec;
 
+use stdClass;
+
 /**
  * This is a simple class codec used for `stdClass` objects.
+ *
  * @internal
- * @todo This class should @implements JsonClassCodec<stdClass> but phan's incomplete
- * support for generics doesn't allow that yet.
+ * @implements JsonClassCodec<stdClass>
  */
 class JsonStdClassCodec implements JsonClassCodec {
 
@@ -35,9 +23,8 @@ class JsonStdClassCodec implements JsonClassCodec {
 	 * can be deserialized with the corresponding newFromJsonArray() method,
 	 * using a ::toJsonArray() method on the object itself.
 	 *
-	 * @param object $obj An object of the type handled by this JsonClassCodec
+	 * @param stdClass $obj An object of the type handled by this JsonClassCodec
 	 * @return array A Json representation of the object.
-	 * @inheritDoc
 	 * @see JsonCodecableTrait
 	 */
 	public function toJsonArray( $obj ): array {
@@ -47,14 +34,12 @@ class JsonStdClassCodec implements JsonClassCodec {
 	/**
 	 * Creates a new instance of the given class and initializes it from the
 	 * $json array, using a static method on $className.
-	 * @param class-string<T> $className
+	 *
+	 * @param class-string<stdClass> $className
 	 * @param array $json
-	 * @return T
-	 * @inheritDoc
-	 * @phan-template T
+	 * @return stdClass
 	 */
 	public function newFromJsonArray( string $className, array $json ) {
-		// @phan-suppress-next-line PhanTypeMismatchReturn inadequate generics
 		return (object)$json;
 	}